+++ frontend/components/inputs/list/Row.js
@@ -3,17 +3,37 @@ import React from 'react';
 export default function Row( {
 	id : key,
 	deleteListRow,
+	addListRow,
+	setRow,
+	row,
 } ) {
 
 	const deleteSelf = () => {
-		deleteListRow( key );
+		if ( deleteListRow ) {
+			deleteListRow( key );
+		} else {
+			alert( "You can't delete a first row" );
+		}
+	};
+
+	const addNextRow = ( e ) => {
+		if ( e.keyCode == 13 ) {
+			if ( addListRow ) {
+				addListRow();
+			} else {
+				const addButton = e.target
+					.closest( '.list' )
+					.querySelector( '.list__add' )
+					.querySelector( 'button' );
+				addButton.click();
+			}
+		}
 	};
 
 	return (
-		<div className='list__row'>
-			<span>Input Item</span>
-			<input></input>
-			{ deleteListRow && <button onClick={deleteSelf}>Delete</button> }
+		<div className={[ 'list__row', `listRow${key}`, ].join( ' ' )}>
+			<input onKeyDown={addNextRow} placeholder="value"></input>
+			<button className="bi bi-trash-fill" onClick={deleteSelf}></button>
 		</div>
 	);
 }
